+++ src/lib/i18n/units.ts
@@ -16,7 +16,7 @@ export type UnitType =
   | 'torque'
   | 'data'
   | 'data_rate'
-  | 'currency'
+  // | 'currency' // 注销货币单位,变化太快
   | 'cooking'
   | 'numeral'
   | 'frequency'